refactor:#265 client onboarding 리팩토링
