+++ src/app/(client)/onboarding/_components/cta-card.tsx
