+++ src/app/(client)/onboarding/_components/first-section.tsx
