+++ src/app/(client)/onboarding/_components/second-section.tsx
