+++ src/app/(client)/onboarding/_components/third-section.tsx
